requirements.txt
absl-py==1.3.0
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
clip==1.0
cycler==0.11.0
Django==3.2.16
django-prefetch==1.2.3
filelock==3.8.0
fonttools==4.37.4
ftfy==6.1.1
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.10.0
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
pip==22.2.2
prefetch==0.0
prefetch-generator==1.0.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
setuptools==63.4.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.6.11
tokenizers==0.12.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.9.0
